+++ variants/views.py
@@ -15,14 +15,16 @@ from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.forms.models import model_to_dict
 from django.http import HttpResponse, Http404, JsonResponse
 from django.db import transaction
+from django.db.models import Q, Sum
 from django.shortcuts import render, redirect, get_object_or_404, reverse
 from django.utils import timezone
 from django.views.generic import DetailView, FormView, ListView, View
 from django.views.generic.detail import SingleObjectMixin, SingleObjectTemplateResponseMixin
+
 import simplejson as json
 
 from bgjobs.models import BackgroundJob
-from clinvar.models import Clinvar
+from clinvar.models import Clinvar, ClinvarPathogenicGenes
 from geneinfo.models import (
     Hgnc,
     NcbiGeneInfo,
@@ -1446,6 +1448,9 @@ class SmallVariantDetails(
             gene["omim"] = omim
             gene["hpo_inheritance"] = list(hpoinheritance)
             gene["hpo_terms"] = list(hpoterms)
+            gene["clinvar_pathogenicity"] = ClinvarPathogenicGenes.objects.filter(
+                entrez_id=gene["entrez_id"]
+            ).first()
             return gene
 
     def _handle_hpo_omim(self, entrez_id):